termino

+++ src/main/java/com/cosanostra/controller/EventoController.java
@@ -63,7 +63,8 @@ public class EventoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminar(@PathVariable Long id) {
+    // --- CORRECCIÓN AQUÍ: Cambiado Long a Integer ---
+    public ResponseEntity<?> eliminar(@PathVariable Integer id) {
         try {
             eventoService.delate(id);
             return ResponseEntity.noContent().build();
