+++ src/main/java/com/cosanostra/repository/SeguridadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeguridadRepository extends JpaRepository<Seguridad, Long> {
+public interface SeguridadRepository extends JpaRepository<Seguridad, Integer> { // <-- ¡Cambio aquí: de Long a Integer!
 }
